requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
cssselect2==0.7.0
diff-match-patch==20230430
Django==5.0.6
django-ckeditor==6.7.1
django-crispy-forms==2.2
django-import-export==4.1.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
docopt==0.6.2
fonttools==4.53.1
html5lib==1.1
idna==3.7
import-export==0.3.1
lxml==5.2.2
mysqlclient==2.2.4
num2words==0.5.13
oscrypto==1.3.0
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyMySQL==1.1.1
pypdf==4.2.0
pyphen==0.15.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.2
reportlab==4.2.2
requests==2.32.3
setuptools==71.0.1
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tablib==3.5.0
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3
